refactor: HumanChatPage 애니메이션 관리 훅으로 분리
- UI와 로직을 분리하기 위해
